+++ src/repository.ts
@@ -6,10 +6,8 @@ import * as cinerino from '@cinerino/domain';
 
 import { MongoRepository as AggregateSaleRepo } from './repo/aggregateSale';
 import { RedisRepository as EventWithAggregationRepo } from './repo/event';
-import { RedisRepository as PaymentNoRepo } from './repo/paymentNo';
 import { MongoRepository as PerformanceRepo } from './repo/performance';
 import { RedisRepository as CheckinGateRepo } from './repo/place/checkinGate';
-import { RedisRepository as TicketTypeCategoryRateLimitRepo } from './repo/rateLimit/ticketTypeCategory';
 import { MongoRepository as ReservationRepo } from './repo/reservation';
 
 /**
@@ -22,24 +20,12 @@ export class AggregateSale extends AggregateSaleRepo { }
  */
 export class EventWithAggregation extends EventWithAggregationRepo { }
 
-export namespace rateLimit {
-    /**
-     * 券種カテゴリーレート制限リポジトリ
-     */
-    export class TicketTypeCategory extends TicketTypeCategoryRateLimitRepo { }
-}
-
 export namespace place {
     /**
      * 入場場所リポジトリ
      */
     export class CheckinGate extends CheckinGateRepo { }
 }
-
-/**
- * 購入番号リポジトリ
- */
-export class PaymentNo extends PaymentNoRepo { }
 
 /**
  * パフォーマンスリポジトリ